Refactor test classes to fix basic setup failures
