+++ integration-tests/test-cases/scenario-tests/src/test/java/io/cellery/integration/scenario/tests/Constants.java
@@ -28,7 +28,7 @@ public class Constants {
     public static final String CELL_ORG_NAME = "myorg";
     public static final String TEST_CELL_ORG_NAME = "myorg";
     public static final String SAMPLE_CELLS_VERSION = "1.0.0";
-    static final String HELLO_WORLD_WEB_CONTENT = "hello cellery";
+    public static final String HELLO_WORLD_WEB_CONTENT = "hello cellery";
     public static final String PET_STORE_WEB_CONTENT = "Pet Accessories";
     public static final String OBSERVABILITY_WEB_CONTENT = "Observability";
     public static final String DEFAULT_HELLO_WORLD_URL = "http://hello-world.com";
