+++ integration-tests/test-cases/scenario-tests/src/test/java/io/cellery/integration/scenario/tests/helloworld/HelloworldWebExtendedTestCase.java
@@ -15,45 +15,32 @@
  * under the License.
  *
  */
-package io.cellery.integration.scenario.tests;
+package io.cellery.integration.scenario.tests.helloworld;
 
+import io.cellery.integration.scenario.tests.ObservabilityDashboard;
 import io.cellery.integration.scenario.tests.models.Cell;
 import io.cellery.integration.scenario.tests.models.SequenceDiagram;
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.io.IOException;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static io.github.bonigarcia.wdm.DriverManagerType.CHROME;
-
 /**
  * This includes the test cases related to hello world web scenario.
  */
-public class HelloworldWebTestCase extends BaseTestCase {
+public class HelloworldWebExtendedTestCase extends HelloworldWebBaseTestCase {
 
     private static final String IMAGE_NAME = "hello-world-web";
     private static final String VERSION = "1.0.0";
     private static final String HELLO_WORLD_INSTANCE = "hello-world-inst";
-    private WebDriver webDriver;
     private ObservabilityDashboard observabilityDashboard;
 
     @BeforeClass
     public void setup() {
-        WebDriverManager.getInstance(CHROME).setup();
-        webDriver = new ChromeDriver(new ChromeOptions().setHeadless(true));
-        observabilityDashboard = new ObservabilityDashboard(webDriver, new WebDriverWait(webDriver, 120));
+        observabilityDashboard = new ObservabilityDashboard(webDriver, webDriverWait);
         observabilityDashboard.setWebCellUrl("http://hello-world.com/");
         List<String> componentList = new ArrayList<>(Arrays.asList("gateway", "hello"));
         Cell cell = new Cell(IMAGE_NAME, HELLO_WORLD_INSTANCE, componentList);
@@ -61,23 +48,6 @@ public class HelloworldWebTestCase extends BaseTestCase {
 
         SequenceDiagram diagram = new SequenceDiagram();
         diagram.getComponents().put(HELLO_WORLD_INSTANCE, 2);
-    }
-
-    @Test
-    public void buildHelloWorld() throws Exception {
-        build("web.bal", Constants.TEST_CELL_ORG_NAME, IMAGE_NAME, VERSION,
-                Paths.get(CELLERY_SCENARIO_TEST_ROOT, "hello-web").toFile().getAbsolutePath());
-    }
-
-    @Test(dependsOnMethods = "buildHelloWorld")
-    public void runHelloWorld() throws Exception {
-        run(Constants.TEST_CELL_ORG_NAME, IMAGE_NAME, VERSION, HELLO_WORLD_INSTANCE, 600);
-    }
-
-    @Test(dependsOnMethods = "runHelloWorld")
-    public void invoke() {
-        webDriver.get(Constants.DEFAULT_HELLO_WORLD_URL);
-        validateWebPage();
     }
 
     @Test (description = "Validates the login flow of the dashboard", dependsOnMethods = "invoke")
@@ -105,33 +75,19 @@ public class HelloworldWebTestCase extends BaseTestCase {
         observabilityDashboard.logoutObservability();
     }
 
+    @Override
     @Test(dependsOnMethods = "observabilityLogout")
     public void terminate() throws Exception {
         terminateCell(HELLO_WORLD_INSTANCE);
     }
 
-    @Test(expectedExceptions = Exception.class, dependsOnMethods = "terminate")
-    public void repeatTerminate() throws Exception {
-        terminateCell(HELLO_WORLD_INSTANCE);
-    }
-
-    @Test(dependsOnMethods = "terminate")
-    public void deleteImage() throws Exception {
-        delete(Constants.TEST_CELL_ORG_NAME + Constants.FORWARD_SLASH + IMAGE_NAME + Constants.COLON + VERSION);
-    }
-
-    private void validateWebPage() {
-        String searchHeader = webDriver.findElement(By.cssSelector("H1")).getText().toLowerCase();
-        Assert.assertEquals(searchHeader, Constants.HELLO_WORLD_WEB_CONTENT, "Web page is content is not as expected");
-    }
-
     @AfterClass
-    public void cleanup() throws InterruptedException, IOException {
-        webDriver.close();
+    public void cleanup() {
+        super.cleanup();
         try {
-            terminateCell(HELLO_WORLD_INSTANCE);
+            observabilityDashboard.cleanupCelleryDashboard();
         } catch (Exception ignored) {
+
         }
-        observabilityDashboard.cleanupCelleryDashboard();
     }
 }
