+++ integration-tests/test-cases/scenario-tests/src/test/resources/basic.xml
@@ -23,24 +23,8 @@
         <classes>
             <class name="io.cellery.integration.scenario.tests.cli.ImageBasedTestCase"/>
             <class name="io.cellery.integration.scenario.tests.cli.InstanceBasedTestCase"/>
-            <class name="io.cellery.integration.scenario.tests.HelloworldWebTestCase">
-                <methods>
-                    <exclude name="observabilityLogin"/>
-                    <exclude name="overviewPage"/>
-                    <exclude name="cellsPage"/>
-                    <exclude name="tracingPage"/>
-                    <exclude name="observabilityLogout"/>
-                </methods>
-            </class>
-            <class name="io.cellery.integration.scenario.tests.petstore.PetStoreTestCase">
-                <methods>
-                    <exclude name="observabilityLogin"/>
-                    <exclude name="overviewPage"/>
-                    <exclude name="cellsPage"/>
-                    <exclude name="tracingPage"/>
-                    <exclude name="observabilityLogout"/>
-                </methods>
-            </class>
+            <class name="io.cellery.integration.scenario.tests.helloworld.HelloworldWebBaseTestCase"/>
+            <class name="io.cellery.integration.scenario.tests.petstore.PetStoreBaseTestCase"/>
         </classes>
     </test>
 </suite>
